Changes in routes and user validation with API
Now, the student is validated using his ID, which is searched for in an external API. If he is found and his profile is that of a student, he is granted access. In addition, the access routes from the frontend have been reorganized. In the backend, the routes are now managed automatically using a foreach loop, instead of being defined manually in the app.js file.